Removed v1 pricing sheet lambda
