+++ lambdas/pricing_sheet/documents/generators.go
@@ -1,0 +1,5 @@
+package documents
+
+type PricingSheetGenerator interface {
+	Generate() (string, error)
+}
