+++ lambdas/pricing_sheet/documents/square_foot.go
@@ -1,0 +1,12 @@
+package documents
+
+type SquareFootPricingSheet struct {
+}
+
+func NewSquareFootPricingSheet() SquareFootPricingSheet {
+	return SquareFootPricingSheet{}
+}
+
+func (p SquareFootPricingSheet) Generate() (string, error) {
+	return "", nil
+}
