+++ lambdas/pricing_sheet/go.mod
@@ -1,19 +1,14 @@
-module app.bluezoneautomation.com/lambda-pricing-sheet
+module app.bluezoneautomation.com/lambda-pricing-sheet/v2
 
 go 1.19
 
 require (
 	github.com/aws/aws-lambda-go v1.41.0
-	github.com/aws/aws-sdk-go v1.45.23
-	github.com/google/uuid v1.3.1
-	github.com/jackc/pgx/v5 v5.4.3
+	github.com/google/uuid v1.4.0
 	github.com/xuri/excelize/v2 v2.8.0
 )
 
 require (
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
 	github.com/richardlehane/mscfb v1.0.4 // indirect
 	github.com/richardlehane/msoleps v1.0.3 // indirect
