--- lambdas/pricing_sheet/utils.go
@@ -1,10 +1,0 @@
-package main
-
-// Convert the boolean to an integer (0/1)
-func BoolToInt(state bool) int {
-	if state {
-		return 1
-	}
-
-	return 0
-}
