--- lambdas/pricing_sheet_v2/documents/generators.go
@@ -1,5 +1,0 @@
-package documents
-
-type PricingSheetGenerator interface {
-	Generate() (string, error)
-}
