--- lambdas/pricing_sheet_v2/documents/square_foot.go
@@ -1,12 +1,0 @@
-package documents
-
-type SquareFootPricingSheet struct {
-}
-
-func NewSquareFootPricingSheet() SquareFootPricingSheet {
-	return SquareFootPricingSheet{}
-}
-
-func (p SquareFootPricingSheet) Generate() (string, error) {
-	return "", nil
-}
