--- lambdas/pricing_sheet_v2/go.mod
@@ -1,20 +1,0 @@
-module app.bluezoneautomation.com/lambda-pricing-sheet/v2
-
-go 1.19
-
-require (
-	github.com/aws/aws-lambda-go v1.41.0
-	github.com/google/uuid v1.4.0
-	github.com/xuri/excelize/v2 v2.8.0
-)
-
-require (
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
-	github.com/richardlehane/mscfb v1.0.4 // indirect
-	github.com/richardlehane/msoleps v1.0.3 // indirect
-	github.com/xuri/efp v0.0.0-20230802181842-ad255f2331ca // indirect
-	github.com/xuri/nfp v0.0.0-20230819163627-dc951e3ffe1a // indirect
-	golang.org/x/crypto v0.12.0 // indirect
-	golang.org/x/net v0.14.0 // indirect
-	golang.org/x/text v0.12.0 // indirect
-)
